PLANNER-2205 optaplanner-integration becomes optaplanner-spring-integration
