--- optaplanner-integration/optaplanner-spring-boot-autoconfigure/src/main/java/org/optaplanner/spring/boot/autoconfigure/SolverManagerProperties.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2019 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.optaplanner.spring.boot.autoconfigure;
-
-public class SolverManagerProperties {
-
-    /**
-     * The number of solvers that run in parallel. This directly influences CPU consumption.
-     * Defaults to "AUTO".
-     * Other options include a number or formula based on the available processor count.
-     */
-    private String parallelSolverCount;
-
-    // ************************************************************************
-    // Getters/setters
-    // ************************************************************************
-
-    public String getParallelSolverCount() {
-        return parallelSolverCount;
-    }
-
-    public void setParallelSolverCount(String parallelSolverCount) {
-        this.parallelSolverCount = parallelSolverCount;
-    }
-
-}
