+++ optaplanner-spring-integration/optaplanner-spring-boot-autoconfigure/.gitignore
