+++ optaplanner-spring-integration/optaplanner-spring-boot-autoconfigure/src/main/java/org/optaplanner/spring/boot/autoconfigure/TerminationProperties.java
@@ -1,0 +1,71 @@
+/*
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.optaplanner.spring.boot.autoconfigure;
+
+import java.time.Duration;
+
+public class TerminationProperties {
+
+    /**
+     * How long the solver can run.
+     * For example: "30s" is 30 seconds. "5m" is 5 minutes. "2h" is 2 hours. "1d" is 1 day.
+     * Also supports ISO-8601 format, see java.time.Duration.
+     */
+    private Duration spentLimit;
+    /**
+     * How long the solver can run without finding a new best solution after finding a new best solution.
+     * For example: "30s" is 30 seconds. "5m" is 5 minutes. "2h" is 2 hours. "1d" is 1 day.
+     * Also supports ISO-8601 format, see java.time.Duration.
+     */
+    private Duration unimprovedSpentLimit;
+    /**
+     * Terminates the solver when a specific or higher score has been reached.
+     * For example: "0hard/-1000soft" terminates when the best score changes from "0hard/-1200soft" to "0hard/-900soft".
+     * Wildcards are supported to replace numbers.
+     * For example: "0hard/*soft" to terminate when any feasible score is reached.
+     */
+    private String bestScoreLimit;
+
+    // ************************************************************************
+    // Getters/setters
+    // ************************************************************************
+
+    public Duration getSpentLimit() {
+        return spentLimit;
+    }
+
+    public void setSpentLimit(Duration spentLimit) {
+        this.spentLimit = spentLimit;
+    }
+
+    public Duration getUnimprovedSpentLimit() {
+        return unimprovedSpentLimit;
+    }
+
+    public void setUnimprovedSpentLimit(Duration unimprovedSpentLimit) {
+        this.unimprovedSpentLimit = unimprovedSpentLimit;
+    }
+
+    public String getBestScoreLimit() {
+        return bestScoreLimit;
+    }
+
+    public void setBestScoreLimit(String bestScoreLimit) {
+        this.bestScoreLimit = bestScoreLimit;
+    }
+
+}
