+++ optaplanner-spring-integration/optaplanner-spring-boot-autoconfigure/src/main/resources/META-INF/spring.factories
