+++ optaplanner-spring-integration/optaplanner-spring-boot-autoconfigure/src/test/resources/org/optaplanner/spring/boot/autoconfigure/customSpringBootSolverConfig.xml
