+++ optaplanner-spring-integration/optaplanner-spring-boot-autoconfigure/src/test/resources/solverConfig.xml
