+++ optaplanner-spring-integration/pom.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.optaplanner</groupId>
+    <artifactId>optaplanner-build-parent</artifactId>
+    <version>8.0.0-SNAPSHOT</version>
+    <relativePath>../optaplanner-build-parent/pom.xml</relativePath>
+  </parent>
+
+  <artifactId>optaplanner-spring-integration</artifactId>
+  <packaging>pom</packaging>
+
+  <name>OptaPlanner Spring integration</name>
+  <description>
+    OptaPlanner solves planning problems.
+    This lightweight, embeddable planning engine implements powerful and scalable algorithms
+    to optimize business resource scheduling and planning.
+
+    This module contains the integration modules.
+  </description>
+  <url>https://www.optaplanner.org</url>
+
+  <properties>
+    <format.directory>${project.basedir}/../../ide-configuration</format.directory>
+  </properties>
+
+  <modules>
+    <module>optaplanner-spring-boot-autoconfigure</module>
+    <module>optaplanner-spring-boot-starter</module>
+  </modules>
+
+</project>
